ts_project/pyproject.toml
[tool.poetry]
name = "ts-project"
version = "0.1.0"
description = ""
authors = ["simonyelisey <yeliseevsemyon@gmail.com>"]
readme = "README.md"
packages = [{include = "ts_project"}]

[tool.poetry.dependencies]
python = "^3.9"
pandas = "^2.1.3"
darts = "^0.27.0"
pylint = "^3.0.2"
black = "22.3.0"
pre-commit = "^3.5.0"
flake8 = "^6.1.0"
hydra-core = "^1.3.2"
numpy = "^1.26.2"
catboost = "^1.2.2"
pyarrow = "^14.0.1"
dvc = "^3.31.1"
dvc-gdrive = "^2.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
